Upgrade to Spring Boot 3 and Wro4j 2, bump to Java 17.
This is a breaking change. If you want to go to Spring Boot 3 with Wro4j, things include the change from Java EE to Jakarta EE, which Wro4j 2 undertakes.
While the general config which this project provides still works and runs fine, we didn’t check the processors and other stuff that come as transitive dependencies and we don’t have plans todo so.